Update mid-level DynamoDB examples.
