+++ dotnetv3/dynamodb/MidLevelBatchWriteItemExample/MidLevelBatchWriteItemExample.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MidLevelBatchWriteItemExample", "MidLevelBatchWriteItemExample\MidLevelBatchWriteItemExample.csproj", "{04DAFEF7-3322-477C-9D67-140E16766A1A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Debug|x64.Build.0 = Debug|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Debug|x86.Build.0 = Debug|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Release|x64.ActiveCfg = Release|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Release|x64.Build.0 = Release|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Release|x86.ActiveCfg = Release|Any CPU
+		{04DAFEF7-3322-477C-9D67-140E16766A1A}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
