--- dotnetv3/dynamodb/MidLevelBatchWriteItemTest/MidLevelBatchWriteItemTest.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AWSSDK.Core" Version="3.5.1.23" />
-    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.5.1.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="xunit" Version="2.4.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
-    <PackageReference Include="coverlet.collector" Version="1.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\CreateTablesLoadData\CreateTablesLoadData.csproj" />
-    <ProjectReference Include="..\MidLevelBatchWriteItem\MidLevelBatchWriteItem.csproj" />
-  </ItemGroup>
-
-</Project>
