--- dotnetv3/dynamodb/MidLevelQueryAndScan/MidLevelQueryAndScan.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AWSSDK.Core" Version="3.5.1.24" />
-    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.5.1.2" />
-  </ItemGroup>
-
-</Project>
