+++ dotnetv3/dynamodb/MidLevelScanOnlyExample/MidLevelScanOnlyExample.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MidLevelScanOnlyExample", "MidLevelScanOnlyExample\MidLevelScanOnlyExample.csproj", "{0680F411-06FB-49D9-8533-0FA8FBBCF56A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Debug|x64.Build.0 = Debug|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Debug|x86.Build.0 = Debug|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Release|x64.ActiveCfg = Release|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Release|x64.Build.0 = Release|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Release|x86.ActiveCfg = Release|Any CPU
+		{0680F411-06FB-49D9-8533-0FA8FBBCF56A}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
